build: limit TS files max lines to 300
