+++ eslint.config.ts
@@ -119,6 +119,16 @@ export default defineConfig(
       eqeqeq: ["error", "always"],
       curly: ["error", "all"],
       "@typescript-eslint/only-throw-error": "error",
+
+      // ── Max lines per file ───────────────────────────
+      "max-lines": [
+        "error",
+        {
+          max: 300,
+          skipBlankLines: true,
+          skipComments: true,
+        },
+      ],
     },
   },
