+++ vitest.config.ts
@@ -19,10 +19,10 @@ export default defineConfig({
       reporter: ["text", "lcov"],
       reportsDirectory: "./test/.coverage",
       thresholds: {
-        statements: 90,
-        branches: 90,
-        functions: 90,
-        lines: 90,
+        statements: 99,
+        branches: 95,
+        functions: 99,
+        lines: 99,
       },
     },
     restoreMocks: true,
